refactor: 장기 연결 레이어 위치 변경
