+++ src/main/java/repository/connect/ConnectionProvider.java
@@ -1,7 +1,9 @@
-package db;
+package repository.connect;
 
 import java.sql.Connection;
 
 public interface ConnectionProvider {
     Connection getConnection();
+
+    void closeConnection(Connection connection);
 }
